requirements.txt
# --- Phase 1: Data Foundation ---
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
psycopg2-binary==2.9.9
faker==24.11.0
sdv==1.14.0
python-dotenv==1.0.1

# --- Phase 2: API & Web Server ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
pydantic==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
pytest==8.1.1
httpx==0.27.0
python-multipart
pytest-asyncio